git-svn-id: https://giordano.mit.edu/nvs/mangle/branches/dev_64bit@127 9f78a35e-8a50-488b-8db1-b351ce30c53d

+++ scripts/graphmasksm.sh
@@ -77,7 +77,9 @@ fi
 if [ -e smerr.temp ] || [ ! -e $2 ]; then
     echo <&2 "ERROR: error in sm plotting script."
     echo <&2 "See log in sm.temp for more details."
-    rm smerr.temp
+    if [ -e smerr.temp ]; then
+	rm smerr.temp
+    fi
     exit 1
 fi
 rm sm.temp
